+++ src/blocks/price-list-child/classes.js
@@ -7,15 +7,7 @@ function PositionClasses( attributes ) {
 	var iconimgStyle_class = ""
 
 	let imge_cnt = 0
-	// attributes.rest_menu_item_arr.map( ( item, thisIndex ) => {
-	// 	let image_arr = attributes.rest_menu_item_arr[thisIndex]
-	// 	if( image_arr && typeof image_arr !== "undefined"){
-	// 		const image = image_arr["image"]
-	// 		if( typeof image !== "undefined" && image !== null && image !=="" ){
-	// 			imge_cnt++
-	// 		}
-	// 	}
-	// } )
+
 	const image = attributes.image
 	if( typeof image !== "undefined" && image !== null && image !=="" ){
 					imge_cnt++
