+++ src/blocks/price-list-child/style.scss
@@ -4,7 +4,7 @@
  * CSS for both Frontend+Backend.
  */
 
-
+ 
 .uagb-rm__desk-column-3 [data-type="uagb/restaurant-menu-child"]{    
     width: calc(100% / 3);   
     display: inline-block;
