+++ src/blocks/price-list/attributes.js
@@ -125,6 +125,161 @@ const attributes = {
 		type: "string",
 		default: "#b2b4b5",
 	},
+	descColor: {
+		type: "string",
+		default: "#333"
+	},
+	priceColor: {
+		type: "string",
+		default: "#888888"
+	},
+	titleColor: {
+		type: "string",
+		default: "#333"
+	},
+	titleFontSizeType: {
+		type: "string",
+		default: "px"
+	},
+	titleFontSize: {
+		type: "number",
+	},
+	titleFontSizeTablet: {
+		type: "number",
+	},
+	titleFontSizeMobile: {
+		type: "number",
+	},
+	titleFontFamily: {
+		type: "string",
+		default: "Default",
+	},
+	titleFontWeight: {
+		type: "string",
+	},
+	titleFontSubset: {
+		type: "string",
+	},
+	titleLineHeightType: {
+		type: "string",
+		default: "em"
+	},
+	titleLineHeight: {
+		type: "number",
+	},
+	titleLineHeightTablet: {
+		type: "number",
+	},
+	titleLineHeightMobile: {
+		type: "number",
+	},
+	titleLoadGoogleFonts: {
+		type: "boolean",
+		default: false	
+	},
+
+	priceFontSizeType: {
+		type: "string",
+		default: "px"
+	},
+	priceFontSize: {
+		type: "number",
+	},
+	priceFontSizeTablet: {
+		type: "number",
+	},
+	priceFontSizeMobile: {
+		type: "number",
+	},
+	priceFontFamily: {
+		type: "string",
+		default: "Default",
+	},
+	priceFontWeight: {
+		type: "string",
+	},
+	priceFontSubset: {
+		type: "string",
+	},
+	priceLineHeightType: {
+		type: "string",
+		default: "em"
+	},
+	priceLineHeight: {
+		type: "number",
+	},
+	priceLineHeightTablet: {
+		type: "number",
+	},
+	priceLineHeightMobile: {
+		type: "number",
+	},
+	priceLoadGoogleFonts: {
+		type: "boolean",
+		default: false	
+	},
+
+	descFontSizeType: {
+		type: "string",
+		default: "px"
+	},
+	descFontSize: {
+		type: "number",
+	},
+	descFontSizeTablet: {
+		type: "number",
+	},
+	descFontSizeMobile: {
+		type: "number",
+	},
+	descFontFamily: {
+		type: "string",
+		default: "Default",
+	},
+	descFontWeight: {
+		type: "string",
+	},
+	descFontSubset: {
+		type: "string",
+	},
+	descLineHeightType: {
+		type: "string",
+		default: "em"
+	},
+	descLineHeight: {
+		type: "number",
+	},
+	descLineHeightTablet: {
+		type: "number",
+	},
+	descLineHeightMobile: {
+		type: "number",
+	},
+	descLoadGoogleFonts: {
+		type: "boolean",
+		default: false	
+	},
+	headingTag: {
+		type: "string",
+		default: "h4"
+	},
+	imagePosition:{
+		type: "string",
+		default: "top"
+	},
+	imageAlignment:{
+		type: "string",
+		default: "top"
+	},
+	imageSize:{
+		type: "string",
+		default: "medium",
+	},
+	imageWidth :{
+		type: "number",
+		default: "",
+	},
+	
 }
 
 export default attributes
