+++ src/blocks/price-list/deprecated.js
