+++ src/blocks/price-list/editor.scss
