+++ src/blocks/price-list/style.scss
